权限中台 任务调度

+++ toucan-shopping-cloud/apps/toucan-shopping-admin-auth-scheduler/src/main/java/com/toucan/shopping/cloud/apps/admin/auth/scheduler/scheduler/AdminRoleToESCacheScheduler.java
@@ -67,50 +67,51 @@ public class AdminRoleToESCacheScheduler {
     @Scheduled(cron = "0 0 0 1/1 * ?")
     public void rerun()
     {
-        logger.info("缓存账号角色信息到ElasticSearch 开始=====================");
-        try {
+        if(toucan.getAdminAuthScheduler().isLoopEsCache()) {
+            logger.info("缓存账号角色信息到ElasticSearch 开始=====================");
+            try {
 
-            //删除索引
-            adminRoleElasticSearchService.deleteIndex();
-
-
-            //如果不存在索引就创建一个
-            while (!adminRoleElasticSearchService.existsIndex()) {
-                adminRoleElasticSearchService.createIndex();
-            }
+                //删除索引
+                adminRoleElasticSearchService.deleteIndex();
 
 
-            int page = 1;
-            int limit = 500;
-            PageInfo pageInfo = null;
-            do {
-                logger.info(" 查询账号角色列表 页码:{} 每页显示 {} ", page, limit);
-                AdminRolePageInfo query = new AdminRolePageInfo();
-                query.setLimit(limit);
-                query.setPage(page);
-                pageInfo = queryPage(query);
-                page++;
-                if (pageInfo != null && org.apache.commons.collections.CollectionUtils.isNotEmpty(pageInfo.getList())) {
-                    String adminRoleListJson = JSONObject.toJSONString(pageInfo.getList());
-                    List<AdminRoleVO> adminRoleVOS = JSONArray.parseArray(adminRoleListJson, AdminRoleVO.class);
+                //如果不存在索引就创建一个
+                while (!adminRoleElasticSearchService.existsIndex()) {
+                    adminRoleElasticSearchService.createIndex();
+                }
 
-                    logger.info("缓存账号角色列表 到Elasticsearch {}", adminRoleListJson);
-                    for (AdminRoleVO adminRoleVO : adminRoleVOS) {
-                        List<AdminRoleElasticSearchVO> adminRoleElasticSearchVOS = adminRoleElasticSearchService.queryById(adminRoleVO.getId());
-                        //如果缓存不存在账号角色将缓存起来
-                        if (org.apache.commons.collections.CollectionUtils.isEmpty(adminRoleElasticSearchVOS)) {
-                            AdminRoleElasticSearchVO adminRoleElasticSearchVO = new AdminRoleElasticSearchVO();
-                            BeanUtils.copyProperties(adminRoleElasticSearchVO, adminRoleVO);
-                            adminRoleElasticSearchService.save(adminRoleElasticSearchVO);
+
+                int page = 1;
+                int limit = 500;
+                PageInfo pageInfo = null;
+                do {
+                    logger.info(" 查询账号角色列表 页码:{} 每页显示 {} ", page, limit);
+                    AdminRolePageInfo query = new AdminRolePageInfo();
+                    query.setLimit(limit);
+                    query.setPage(page);
+                    pageInfo = queryPage(query);
+                    page++;
+                    if (pageInfo != null && org.apache.commons.collections.CollectionUtils.isNotEmpty(pageInfo.getList())) {
+                        String adminRoleListJson = JSONObject.toJSONString(pageInfo.getList());
+                        List<AdminRoleVO> adminRoleVOS = JSONArray.parseArray(adminRoleListJson, AdminRoleVO.class);
+
+                        logger.info("缓存账号角色列表 到Elasticsearch {}", adminRoleListJson);
+                        for (AdminRoleVO adminRoleVO : adminRoleVOS) {
+                            List<AdminRoleElasticSearchVO> adminRoleElasticSearchVOS = adminRoleElasticSearchService.queryById(adminRoleVO.getId());
+                            //如果缓存不存在账号角色将缓存起来
+                            if (org.apache.commons.collections.CollectionUtils.isEmpty(adminRoleElasticSearchVOS)) {
+                                AdminRoleElasticSearchVO adminRoleElasticSearchVO = new AdminRoleElasticSearchVO();
+                                BeanUtils.copyProperties(adminRoleElasticSearchVO, adminRoleVO);
+                                adminRoleElasticSearchService.save(adminRoleElasticSearchVO);
+                            }
                         }
                     }
-                }
-            } while (pageInfo != null && CollectionUtils.isNotEmpty(pageInfo.getList()));
-        }catch(Exception e)
-        {
-            logger.warn(e.getMessage(),e);
+                } while (pageInfo != null && CollectionUtils.isNotEmpty(pageInfo.getList()));
+            } catch (Exception e) {
+                logger.warn(e.getMessage(), e);
+            }
+            logger.info("缓存账号角色信息到ElasticSearch 结束=====================");
         }
-        logger.info("缓存账号角色信息到ElasticSearch 结束=====================");
     }
 
 }
